Adjust test. Refactoring.
