+++ library/src/main/java/bio/terra/landingzone/library/landingzones/definition/factories/ProtectedDataStepsDefinitionProvider.java
@@ -6,7 +6,6 @@ import bio.terra.landingzone.library.landingzones.definition.ArmManagers;
 import bio.terra.landingzone.library.landingzones.definition.DefinitionHeader;
 import bio.terra.landingzone.library.landingzones.definition.DefinitionVersion;
 import bio.terra.landingzone.stairway.flight.ResourceNameProvider;
-import bio.terra.landingzone.stairway.flight.StepsDefinitionFactoryType;
 import bio.terra.landingzone.stairway.flight.create.resource.step.ConnectLongTermLogStorageStep;
 import bio.terra.landingzone.stairway.flight.create.resource.step.CreateAksLogSettingsStep;
 import bio.terra.landingzone.stairway.flight.create.resource.step.CreateSentinelAlertRulesStep;
@@ -22,7 +21,10 @@ import org.apache.commons.lang3.tuple.Pair;
 
 public class ProtectedDataStepsDefinitionProvider extends CromwellStepsDefinitionProvider {
   private final String LZ_NAME = "Protected data Landing Zone Base Resources";
-  private final String LZ_DESC = "Protected data landing zone description";
+  private final String LZ_DESC =
+      "Protected data landing zones are intended to fulfill legal requirements for data governed by a compliance standard, "
+          + "such as HIPAA protected data, federal controlled-access data, etc. They deploy additional resources to the landing "
+          + "zone for additional security monitoring (using Azure Sentinel) and exporting logs to centralized long-term storage for retention.";
 
   @Override
   public String landingZoneDefinition() {
