+++ library/src/test/java/bio/terra/landingzone/stairway/flight/StepsDefinitionFactoryTypeTest.java
@@ -3,6 +3,7 @@ package bio.terra.landingzone.stairway.flight;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
+import bio.terra.landingzone.library.landingzones.definition.factories.StepsDefinitionFactoryType;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.params.ParameterizedTest;
